Fetch the newest main config for a new release branch instead of all
Avoid replicating all the main config files to a new release branch, and use only the latest which should be aligned OCP version for this particluar release branch. Later on this release branch should include the next OCP version and the next next OCP version if it is even

+++ ci-operator/config/medik8s/common/new-release.sh
@@ -32,18 +32,28 @@ fi
 
 cd ${REPO}
 
-# copy all main config files
-for file in $(ls | grep main__); do
-  # rename the config file with the branch name
-  new_file=$(echo ${file} | sed "s/main__/${BRANCH}__/")
-  cp ${file} ${new_file}
-  # update the branch name in the config file
-  sed -i "s/branch: main/branch: ${BRANCH}/g" ${new_file}
-  # update the old released operator version in the config file if possible
-  if [ $# -eq 3 ]; then
-    sed -i "s/OPERATOR_RELEASED_VERSION: .*/OPERATOR_RELEASED_VERSION: ${OPERATOR_RELEASED_VERSION}/g" ${new_file}
-  fi
-done
+# copy only the newest main config (highest OCP version)
+latest_main=$(ls | grep 'main__' | sed -r 's#^.*__##; s#\\.yaml$##' | sort -V | tail -1)
+if [ -z "${latest_main}" ]; then
+  echo "No main__*.yaml config files found"
+  exit 1
+fi
+
+file="medik8s-${1}-main__${latest_main}"
+if [ ! -f "${file}" ]; then
+  echo "Expected config ${file} not found"
+  exit 1
+fi
+
+# rename the config file with the branch name
+new_file=$(echo ${file} | sed "s/main__/${BRANCH}__/")
+cp ${file} ${new_file}
+# update the branch name in the config file
+sed -i "s/branch: main/branch: ${BRANCH}/g" ${new_file}
+# update the old released operator version in the config file if possible
+if [ $# -eq 3 ]; then
+  sed -i "s/OPERATOR_RELEASED_VERSION: .*/OPERATOR_RELEASED_VERSION: ${OPERATOR_RELEASED_VERSION}/g" ${new_file}
+fi
 
 echo "Done, please run 'make update' for creating jobs"
